feat: Android scale issues and imageQuality option (#716)
## Summary
- Scale handling changes for session replay (scale as Double), including
adapter updates required to compile (Flutter LDNativeApiImpl, React
Native SessionReplayClientAdapter).
- Add ReplayOptions.imageQuality (0.0–1.0, default 0.3), threaded
through CaptureManager into ExportDiffManager to drive JPEG quality.

## Test plan
- ExportDiffManagerTest / ReplayOptionsTest updated; run the
observability-android unit tests.

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Overview**
> Android session replay now treats **replay scale** as `Double`
end-to-end (`ReplayOptions`, capture, export frames, touch scaling) so
the exporter and capture path share the same factor—avoiding oversized
frames when Flutter/RN adapters previously forced `Float` conversions.
Flutter `LDNativeApiImpl` and React Native `SessionReplayClientAdapter`
are updated to pass scale without narrowing to float.
> 
> A new **`ReplayOptions.imageQuality`** setting (0.0–1.0, default
**0.3**) controls JPEG compression when frames are exported; it flows
from `CaptureManager` into `ExportDiffManager` instead of the fixed
default JPEG quality. The Android README documents the option; unit
tests cover defaults, builder configuration, and that encoding uses the
configured quality.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
ec432700dc8ed837c65d7c57472771787327a367. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

---------

+++ sdk/@launchdarkly/flutter/packages/observability/android/src/main/kotlin/com/launchdarkly/launchdarkly_flutter_observability/LDNativeApiImpl.kt
@@ -130,7 +130,7 @@ internal class LDNativeApiImpl(
             enabled = replay.isEnabled ?: true,
             sampleRate = replay.sampleRate ?: 1.0,
             frameRate = replay.frameRate ?: 1.0,
-            scale = replayScale.toFloat(),
+            scale = replayScale,
             privacyProfile = PrivacyProfile(
                 maskTextInputs = maskTextInputs,
                 maskText = privacy?.maskLabels ?: false,
